Split test material across smaller files [BF-2051]
[BF-2051]
