+++ test/helpers/__init__.py
@@ -1,0 +1,1 @@
+# Copyright (c) 2023 Aiven, Helsinki, Finland. https://aiven.io/
